+++ src/lib/components/modal/code/RemoveCode.tsx
@@ -27,6 +27,7 @@ export function RemoveCodeModal({
   name,
   trigger = (
     <StyledIconButton
+      aria-label="button"
       icon={<CustomIcon name="delete" />}
       variant="ghost-gray"
     />