+++ src/lib/pages/account-details/components/modules/ModuleListsBody.tsx
@@ -25,9 +25,8 @@ export const ModuleListsBody = ({
   const filteredModules = useMemo(() => {
     if (!keyword) return modules;
 
-    return modules?.filter(
-      (module) =>
-        module.moduleName?.toLowerCase().includes(keyword.toLowerCase())
+    return modules?.filter((module) =>
+      module.moduleName?.toLowerCase().includes(keyword.toLowerCase())
     );
   }, [keyword, modules]);